Improve validations

+++ app/controllers/api/v1/validation_violations_controller.rb
@@ -4,6 +4,10 @@ class Api::V1::ValidationViolationsController < Api::V1::ApiController
 
     project = current_user.projects.find(params[:project_id])
 
+    unless feature_enabled?(project, Organization::FEATURE_VALIDATIONS)
+      return
+    end
+
     page = parse_page(params[:page])
     per_page = parse_per_page(params[:per_page])
 
@@ -41,11 +45,20 @@ class Api::V1::ValidationViolationsController < Api::V1::ApiController
 
     project = current_user.projects.find(params[:project_id])
 
+    unless feature_enabled?(project, Organization::FEATURE_VALIDATIONS)
+      return
+    end
+
     render json: { total: project.issues_count }
   end
 
   def destroy
     project = current_user.projects.find(params[:project_id])
+
+    unless feature_enabled?(project, Organization::FEATURE_VALIDATIONS)
+      return
+    end
+
     validation_violation = project.validation_violations.find(params[:id])
     authorize validation_violation
     validation_violation.destroy
@@ -55,6 +68,11 @@ class Api::V1::ValidationViolationsController < Api::V1::ApiController
 
   def destroy_multiple
     project = current_user.projects.find(params[:project_id])
+
+    unless feature_enabled?(project, Organization::FEATURE_VALIDATIONS)
+      return
+    end
+
     validation_violations_to_destroy = project.validation_violations.find(params[:validation_violation_ids])
     validation_violations_to_destroy.each { |validation_violation| authorize validation_violation }
     project.validation_violations.destroy(validation_violations_to_destroy)
@@ -64,6 +82,11 @@ class Api::V1::ValidationViolationsController < Api::V1::ApiController
 
   def update
     project = current_user.projects.find(params[:project_id])
+
+    unless feature_enabled?(project, Organization::FEATURE_VALIDATIONS)
+      return
+    end
+
     validation_violation = project.validation_violations.find(params[:id])
     authorize validation_violation
     validation_violation.ignored = params[:ignored]
@@ -74,6 +97,11 @@ class Api::V1::ValidationViolationsController < Api::V1::ApiController
 
   def update_multiple
     project = current_user.projects.find(params[:project_id])
+
+    unless feature_enabled?(project, Organization::FEATURE_VALIDATIONS)
+      return
+    end
+
     validation_violations_to_ignore = project.validation_violations.find(params[:validation_violation_ids])
     validation_violations_to_ignore.each do |validation_violation|
       authorize validation_violation
